+++ CHANGELOG.md
@@ -11,12 +11,16 @@ Antonio Ramirez.
 ## YiiBooster latest development alpha
 - **(enh)** upgrade to select2 3.5.1 and select2-bootstrap-css 1.4.1
 - **(enh)** upgrade to font-awesome 4.2
+- **(fix)** fix TbButtonGroupColumn doesn't conform to bootstrap 3+ - #977 
+- **(fix)** fix Userdoc for select2 still references 'bootstrap.widgets'
+- **(enh)** Added support to TbActiveForm to use TbChosen widget (YaroslavMolchan)
   
 ## YiiBooster version 4.1.0
 - **(enh)** upgarde to Bootstrap 3.2.0 - #882
 - **(enh)** upgarde to bootstrap-switch 3.0.2 - #867
 - **(enh)** add callbacks support to TbFileUpload
 - **(enh)** update TbHtml to the latest 1.3.0, and addeing its dependency TbArray  
+- **(enh)** update bootstrap-datepicker to the latest 1.3.1 (fixes #987, #873)
 - **(fix)** TbActiveForm fix "[a]attr" format attribute #856 
 - **(fix)** allow multi relational column - #872  
 - **(fix)** fix TbButton dropdown items -maybe removed in a future release-